requirements.txt
openai
pinecone-client
langchain
fastapi
uvicorn
python-dotenv
pymongo
streamlit
tiktoken